bees: handle --version and --help/-h before bees initialisation
Introduce the standard --version option to print version and exit 0.
Adjust --help/-h to do the same.

Scan argv in main() for standard options before calling bees_main().
This avoids going through config parsing and Chatter log setup that we're
going to throw away immediately.

If neither option appears, write the version on stderr as before.
Since the version string may be critical to debugging the configuration
parsing code, the version string output is unconditional, and occurs
prior to the start of config parsing.

Resolves: https://github.com/Zygo/bees/issues/330
Signed-off-by: Zygo Blaxell <bees@furryterror.org>